New crypto.uniform(n) function samples uniformly from {0,1,2,...,n-1}.
Automatic tests in test/uniformTest.js are stochastic with nonzero
probability of spurious failure -- i.e., when the sampler is working
correctly, there is a nonzero probability that it will fail tests.
This probability is currently under .5%, i.e. fewer than one in two
hundred runs of the test suite should fail.  If we add more tests,
this number will go up.  With extra computation, we can keep it down.