Eliminate tests for low-precision sampler.
The goal of this was to show that the statistical test we use has the
power to distinguish the naive uniform(n)/n algorithm from the one we
use, in the reduced precision of binary16, with 11 bits of precision.

However, it turns out that this test lacks the statistical power to
distinguish uniform_01() in binary16 from uniform(2^10)/2^10 in
binary16 floating-point with adequate probability for our test runs.

Also, the nonuniformity of floor(uniform_01()*100) in binary16 is
enough to raise the spurious failure rate beyond what is acceptable
here.  To remedy this would require precisely calculating that
distribution, which is doable but not worth the trouble for now.